# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RoleRadar / JobEngine API",
    "description": "v1: sponsor-ranked job listings + per-user application tracker. v2 (/api/v2): evidence-grounded 'should I apply?' decisions with cited resume evidence, plus outcome tracking and ranking evaluation.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/profile": {
      "get": {
        "tags": [
          "roleradar"
        ],
        "summary": "Get Profile",
        "operationId": "get_profile_api_v2_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "roleradar"
        ],
        "summary": "Save Profile",
        "operationId": "save_profile_api_v2_profile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/evidence": {
      "get": {
        "tags": [
          "roleradar"
        ],
        "summary": "List Evidence",
        "description": "Every citable unit extracted from the stored resume.\n\nExposed because citations are only trustworthy if the cited corpus is\ninspectable: a user must be able to see exactly what the scorer can see.",
        "operationId": "list_evidence_api_v2_evidence_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response List Evidence Api V2 Evidence Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/decide": {
      "post": {
        "tags": [
          "roleradar"
        ],
        "summary": "Decide",
        "description": "Score one posting. Anonymous when the request carries its own resume.\n\n`save` and looking a job up by `job_key` both need an account; an inline\njob + inline resume needs nothing.",
        "operationId": "decide_api_v2_decide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Decide Api V2 Decide Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/rank": {
      "post": {
        "tags": [
          "roleradar"
        ],
        "summary": "Rank",
        "description": "Batch-score the listing set for the signed-in user.\n\nNever calls an LLM, no matter what: this path can touch 500 postings, and\nthe explanations shown here are the deterministic ones. A user who opens\na specific role gets the model-written version from `/decide`.",
        "operationId": "rank_api_v2_rank_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rank Api V2 Rank Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/decisions": {
      "get": {
        "tags": [
          "roleradar"
        ],
        "summary": "List Decisions",
        "operationId": "list_decisions_api_v2_decisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Decisions Api V2 Decisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/outcomes": {
      "post": {
        "tags": [
          "roleradar"
        ],
        "summary": "Add Outcome",
        "description": "Legacy convenience route: log a stage for a job_key directly, with no\nAPPLY decision required first. Kept beside the new /jobs/.../decisions +\n/applications flow for CLI and quick-logging callers; both write to the\nsame rr_applications/rr_application_events tables underneath.",
        "operationId": "add_outcome_api_v2_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Add Outcome Api V2 Outcomes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v2/outcomes/funnel": {
      "get": {
        "tags": [
          "roleradar"
        ],
        "summary": "Funnel",
        "operationId": "funnel_api_v2_outcomes_funnel_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Funnel Api V2 Outcomes Funnel Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v2/decisions/{job_decision_id}": {
      "patch": {
        "tags": [
          "roleradar"
        ],
        "summary": "Patch Job Decision",
        "operationId": "patch_job_decision_api_v2_decisions__job_decision_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_decision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Decision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobDecisionPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Job Decision Api V2 Decisions  Job Decision Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/evaluation": {
      "get": {
        "tags": [
          "roleradar"
        ],
        "summary": "Evaluation",
        "description": "Did the ranking actually predict outcomes? Reported honestly, including\nwhen the answer is \"not enough data\" or \"no better than random\".",
        "operationId": "evaluation_api_v2_evaluation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "min_sample",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Min Sample"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Evaluation Api V2 Evaluation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecideIn": {
        "properties": {
          "job_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Key"
          },
          "job": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job"
          },
          "jd_text": {
            "type": "string",
            "maxLength": 50000,
            "title": "Jd Text",
            "default": ""
          },
          "resume": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume"
          },
          "profile": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "use_llm": {
            "type": "boolean",
            "title": "Use Llm",
            "default": false
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "default": false
          }
        },
        "type": "object",
        "title": "DecideIn",
        "description": "Either `job_key` (look the posting up) or an inline `job` dict."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobDecisionPatch": {
        "properties": {
          "decision": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "apply",
                  "save",
                  "skip",
                  "defer"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "JobDecisionPatch"
      },
...
      "OutcomeIn": {
        "properties": {
          "job_key": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Job Key"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "job_key",
          "stage"
        ],
        "title": "OutcomeIn"
      },
      "ProfileIn": {
        "properties": {
          "resume": {
            "additionalProperties": true,
            "type": "object",
            "title": "Resume",
            "description": "resume.json contents"
          },
          "needs_sponsorship": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs Sponsorship"
          },
          "graduation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "us_citizen": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Us Citizen"
          },
          "open_to_relocate": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open To Relocate"
          },
          "seniority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seniority"
          }
        },
        "type": "object",
        "required": [
          "resume"
        ],
        "title": "ProfileIn"
      },
      "ProfileOut": {
        "properties": {
          "profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Profile"
          },
          "evidence_units": {
            "type": "integer",
            "title": "Evidence Units"
          },
          "skills_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Detected"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "evidence_units",
          "skills_detected"
        ],
        "title": "ProfileOut"
      },
      "RankIn": {
        "properties": {
          "grad": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "intern",
                  "newgrad"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Grad"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "title_contains": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Contains"
          },
          "tiers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tiers"
          },
          "scan": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Scan",
            "default": 200
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          },
          "save": {
            "type": "boolean",
            "title": "Save",
            "default": false
          }
        },
        "type": "object",
        "title": "RankIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}